Refactor in `TeamKernelFunctions` folder.
